pyproject.toml
[project]
name = "WorkDataHub"
version = "0.1.0"
description = "A modernized data processing platform for WorkDataHub."
authors = [{ name = "Cline", email = "cline@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "dagster",
    "dagster-webserver",
    "pandas",
    "pandera>=0.18.0,<1.0",
    "pydantic>=2.11.7",
    "structlog",
    "psycopg2-binary",
    "alembic",
    "sqlalchemy>=2.0",
    "openpyxl",
    "PyYAML",
    "pydantic-settings>=2.10.1",
    "playwright>=1.55.0",
    "playwright-stealth>=2.0.0",
    "opencv-python-headless>=4.10.0.84",
    "numpy>=2.0.0",
    "gmssl>=3.2.1",
    "pyarrow>=21.0.0",
    "psutil>=7.0.0",
    "requests>=2.32.0",
    "sqlglot>=28.1.0",
    "PyMySQL>=1.1.0",
    "psycopg2>=2.9.11",
]
 
[tool.setuptools.packages.find]
where = ["src"]
 
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-postgresql",
    "ruff>=0.12.12",
    "mypy>=1.17.1",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "PLR"]
extend-select = ["TID"]

[tool.ruff.lint.per-file-ignores]
"src/work_data_hub/io/**/*.py" = ["TID251"]
"src/work_data_hub/orchestration/**/*.py" = ["TID251"]
"src/work_data_hub/domain/annuity_performance/*.py" = ["E501"]
"src/work_data_hub/scripts/**/*.py" = ["TID251"]
"scripts/**/*.py" = ["TID251"]
"tests/**/*.py" = ["ALL"]
"legacy/**/*.py" = ["ALL"]

[tool.ruff.lint.flake8-tidy-imports.banned-api."work_data_hub.io"]
msg = "Domain layer cannot import from work_data_hub.io (Story 1.6 Clean Architecture)."

[tool.ruff.lint.flake8-tidy-imports.banned-api."work_data_hub.orchestration"]
msg = "Domain layer cannot import from work_data_hub.orchestration (Story 1.6 Clean Architecture)."

[tool.ruff.lint.pylint]
max-statements = 50  # Per function limit (aligns with project-context.md)
max-branches = 12    # Cyclomatic complexity threshold

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [
    ".git",
    ".venv",
    "node_modules",
]
markers = [
    "unit: marks tests as fast unit tests with no external dependencies",
    "integration: marks tests as integration tests with database or filesystem",
    "postgres: marks tests as requiring a PostgreSQL database (deselect with '-m \"not postgres\"')",
    "eqc_integration: marks tests that call the EQC API (opt-in)",
    "monthly_data: marks tests requiring reference/monthly data (opt-in)",
    "legacy_data: marks tests as requiring legacy sample data (opt-in)",
    "e2e: legacy marker for parity/end-to-end scenarios (see e2e_suite)",
    "performance: marks slow or resource-intensive scenarios",
    "legacy_suite: opt-in regression tests that exercise the legacy annuity stack (RUN_LEGACY_TESTS=1 or --run-legacy-tests)",
    "e2e_suite: opt-in Dagster/warehouse E2E flows that require heavy fixtures (RUN_E2E_TESTS=1 or --run-e2e-tests)",
    "sandbox_domain: marks tests for the non-production sandbox domain (sandbox_trustee_performance)"
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pandas-stubs>=2.3.2.250827",
    "pre-commit>=3.5.0",
    "psycopg2-binary>=2.9.10",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.12",
    "tomli>=2.0.0; python_version < '3.11'",
    "types-psutil>=7.0.0.20251115",
    "types-psycopg2>=2.9.21.20250809",
    "types-python-dateutil>=2.9.0.20250822",
    "types-pyyaml>=6.0.12.20250822",
    "types-requests>=2.32.4.20250913",
]

[tool.mypy]
# 让 mypy 将 `src` 视为包根并避免重复模块名
mypy_path = ["src"]
explicit_package_bases = true
namespace_packages = true

# Strict mode settings (Story 1.2 - AC-2)
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
no_implicit_reexport = true
strict_equality = true

# Handle third-party libraries without type stubs
[[tool.mypy.overrides]]
module = [
    "playwright_stealth",
    "cv2",
    "gmssl.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "legacy.*"
ignore_errors = true
